[DELETE] #227 - 안 쓰이는 파일들 삭제
