--- Smeem-iOS/Smeem-iOS/Global/Protocols/DataSend/AuthDataSendProtocol.swift
@@ -1,12 +1,0 @@
-//
-//  AuthDataSendProtocol.swift
-//  Smeem-iOS
-//
-//  Created by 황찬미 on 2023/10/10.
-//
-
-import Foundation
-
-protocol AuthDataSendProtocol {
-    func sendTrainingAlarm()
-}
