--- Smeem-iOS/Smeem-iOS/Global/Protocols/DataSend/DataBindProtocol.swift
@@ -1,12 +1,0 @@
-//
-//  DataBindProtocol.swift
-//  Smeem-iOS
-//
-//  Created by Joon Baek on 2023/06/28.
-//
-
-import Foundation
-
-protocol DataBindProtocol: AnyObject {
-    func dataBind(topicID: Int?, inputText: String)
-}
