+++ Smeem-iOS/Smeem-iOS/Global/Protocols/Store.swift
