+++ Smeem-iOS/Smeem-iOS/Network/NetworkDataModel/Badge/BadgeNetworkModel.swift
