--- Smeem-iOS/Smeem-iOS/Presentation/Onboarding/Training/Way/Model/TrainingWayAppData.swift
@@ -1,16 +1,0 @@
-//
-//  TrainingWayAppData.swift
-//  Smeem-iOS
-//
-//  Created by 황찬미 on 2024/02/07.
-//
-
-import Foundation
-
-struct TrainingWayAppData {
-    let wayTitle: String
-    let wayOne: String
-    let wayTwo: String
-    let detailWayOne: String
-    let detailWayTwo: String
-}
